More modular migration framework.
Only implemented for postgresql, can be ported to sqlite. I've removed
the initialize method for PersistBackend, realizing the initialization
is too backend-specific to be generalized. Now there's a migrate
function which determines which alterations must be made to a database,
a Migrate type synonym to keep the list of alterations, parseMigration
to unwrap the Migrate datatype and runMigration to perform the
migrations.

Future work: we could provide some alternate functions to do other
things with the list of alterations, such as saving them to a file to be
reviewed and executed by the user. dbmigrations package may be something
to consider, though it seems to have issues currently with GHC 6.12.