persistent-mysql: Immediately fetch all results into memory.
Since we're already using 'storeResult' (i.e., mysql_store_result),
all the rows are going to be on the client's memory anyway, but inside
a MySQL buffer.  This commit changes withStmt' so that we fetch to all
rows to Haskell-land as soon as possible and then free the MySQL buffer.

The hugest benefit of this commit is that now you'll be able to use
nested transactions with persistent-mysql.  For example, now you may
use deleteCascadeWhere.

The drawback is that we may end up converting some rows that may not
be needed by the Haskell code.  But since most of the persistent code
out there already uses list variants which will fetch all results
right away anyway, not much has been lost.