Remove `(200)` fixed key length
This resolves the issues mentioned in #330.

according to http://dev.mysql.com/doc/refman/5.1/en/alter-table.html, the syntax is:

```
ADD [CONSTRAINT [symbol]]
        UNIQUE [INDEX|KEY] [index_name]
        [index_type] (index_col_name,...) [index_option] ...
```
where
```
index_col_name:
    col_name [(length)] [ASC | DESC]
```

The `(200)` above, is the `length` part, and therefore will likely only work on fields that are at least `200` long. I suggest we remove the fixed length, until we can properly use the `maxlen` field of the appropriate column, if needed at all.
But as it is, it breaks using Uniques in Yesod. Say:

```
Table
  name    Text maxlen=45 sqltype=varchar(45)
  TableName name
```
would fail.