Fix non-idempotent migration of domains
Postgres supports the concept of domains, which are data types with optional constraints.
An app might make an "email" domain over the varchar type, with a CHECK that the emails are valid
In this case the generated SQL should use the domain name: ALTER TABLE users ALTER COLUMN foo TYPE email

Previously the migration code only looked at the underlying type (varchar), and tried to migrate something already of the domain email back to an email.