added: PersistentTest: upsert fail, it expected equivalent repsert
I see
[haskell - Persistent `upsert` isn't working -
Stack Overflow](https://stackoverflow.com/questions/50050431/persistent-upsert-isnt-working)

The question is that if the second update argument of upsert is empty, the data will not be updated.

Certainly it is written in the document comment as follows.

> leaving this empty is the equivalent of performing a repsert on a unique key
>
> [Database.Persist.Class](https://www.stackage.org/haddock/lts-11.15/persistent-2.8.2/Database-Persist-Class.html#v:upsert)

According to the document, upsert needs to behave like repsert if the second argument is empty.

However, looking at the source code, it seems that such behavior is not supposed to be.

So I wrote a test that failed.

To send test results, this bug report uses pull request instead of issue.

I do not know which is better for me, whether to modify the document or modify the implementation.