Atomic  insertUnique in SqlBackend and PostgreSQL 9.5+ support
The `insertUnique` default implementation first checks for conflicts,
one query per unique constraint, and no conflict is found, it inserts a
new record. That insertion may fail, if concurrent insertions happen
while those conflict checks are being executed.

In PostgreSQL, the default isolation level allows such concurrent
insertions to be visible and make the insertion query throw an error.
But since version 9.5, PostgreSQL has upsert support, which includes the
"ON CONFLICT DO NOTHING" clause. It allows to atomically
insert-only-if-no-conflicts, even in committed-read isolation level,
without needing to switch to repeatable-read.

This commit adds an SqlBackend field allowing SQL backends to provide
SQL text for insert-if-no-conflicts, and SqlBackend's implementation of
insertUnique uses it if provided, otherwise resorting to the old
default.

This commit also provides this new field for the PostgreSQL backend.

Note: The insertion isn't 100% atomic as in actual repeatable-read. For
example, if the inserted record contains a foreign key referring to a
record being concurrently deleted, it will cause an error. But that's
true for regular insert too. The guarantee is that if an error is
raised, unique constraint conflicts aren't the cause, even in a
concurrent situation (while by default, in committed-read it's only
guaranteed for non-concurrent cases).