Expose the knot tying logic of `parseReferences` directly
TH code generation makes up a significant amount of compile time (see also
#924) and forcing all entities that reference each other to be in a single TH
block means that ALL the code generation has to be redone and recompiled
whenever a single entity changes.

The reason that persistent requires all entities to be in the same TH block is
to properly setup the cross-references from foreign keys when generating a
Migration.

All the logic to build a Migration from a list of `EntityDef`s is already
exposed in persistent/persistent-template, so it's perfectly possible to
generate the `Migration` at runtime from independent entity definitions. The
only thing stopping users from doing this is that the logic to setup the
cross-references for foreign keys is...complicated.

`parseReferences` currently fixes these cross-references for entities in a
single block AND generates the resulting TH. This commit simply moves the logic
for fixing the cross-references into a separate, exported function and has
`parseReferences` call that function.

The result is that users can now define entities independently (even in
separate modules), use this new function to fix up the cross-references, and
then generate the resulting `Migration` at runtime.

I've been using this for a while now, defining each entity (and associated
code) in its own module, with no problems and it has substantially reduced the
(re)compile times after schema changes.