Fix multiple fk issue with PostgreSQL migrations (#1010)
* [postgres] Fix multiple foreign keys on a single column

This PR addresses issue #1009 for the PostgreSQL backend. In order to support multiple foreign keys on a single column, we now pass the reference name to getColumn, and use it in the where clause when querying foreign keys for the column.

Adds failing tests for this, and gives a better error message if someone runs into the error.