Add proper support for Word64
Previously when the schema used `Word64` as the column type, Persistent
would use `SqlInt64` as the SQL representation which means that `Word64`
values above `maxBound :: Int64` would be stored as negative values in
the database. That is fine for a database only accessed from Haskell but
is a pain in the neck when the database is used as an interop layer for
other languages.

This commit fixes these issues by adding `SqlWord64` and `PersistWord64`.

Closes: https://github.com/yesodweb/persistent/issues/1095