Simplify type check
For a composite reference, the foreign fields have to be checked against
the parent fields for type equality. The parent field are found by
searching the parent Entity, using 'getFD'. For the foreign
fields,though, this is unecessary. Currently, for each composite field, we extract
its fields name and search again for the same field using 'getFd'. This
creates an unecessary round trip.