Use explicit ordering in RawSqlTest
I was non-deterministically getting an error on this test, where the pets were misordered. I'm having trouble reproducing, but I believe the issue is that Postgres' order is non-deterministic if not specified (in my experience, it can be Postgres version dependent). I believe the test as-is is incorrect without specifying the pet order, for Postgres.

[Postgres docs warn:](https://www.postgresql.org/docs/12/queries-order.html)

> After a query has produced an output table (after the select list has been processed) it can optionally be sorted. If sorting is not chosen, the rows will be returned in an unspecified order. The actual order in that case will depend on the scan and join plan types and the order on disk, but it must not be relied on. A particular output ordering can only be guaranteed if the sort step is explicitly chosen.

This commit orders the pets explicitly. This ensures when we get both of Mathias' pets, we get "Rodolfo" first and "Zeno" second, as the assertions require.