Improve documentation around migration safety
The Haddocks around CautiousMigration were ambiguous, since it wasn't
clear if the flag signified whether each migration was safe or not - and
there is at least one instance in the code that gets it wrong.

This corrects that error, opting to make the `addMigration`
documentation match the actual behaviour of the code.