Add API for mem-constant DB streams, Postgres implementation
This code is based on that contained in:
https://github.com/yesodweb/persistent/pull/726

This commit has two components:

  1. A new backend class, `PersistQueryStream`, for backends which *may*
allow for constant-memory streaming of query results. This is intended
to replace the use of `selectSource` from `PersistQueryRead`, which is
"broken" in the sense that the `ConduitT` it returns rarely actually
does any streaming.
  2. An implementation of `PersistQueryStream` for `SqlBackend` which is
based on an internal field, `connPrepareCursor`, which allows for
backends to provide a way to run queries using an internal cursor which
iterates over rows as requested, instead of returning them all at once.
This internal field is provided for the Postgres version of
`SqlBackend`, using functions from `postgresql-simple`.

This code differs in an important way from the above PR in that it
solves a memory issue in the way the `ConduitT` was constructed that
meant that the implementation using cursors was also not
memory-constant. This implementation is memory-constant, tested to large
sets of results.

Co-authored-by: Tim Stapels <tim@supercede.com>