Add RawPostgresql backend which exposes connection
As part of the discussion in yesodweb/persistent#1298, it was found that
external libraries which want to hook into Persistent's Postgres API are
limited by their inability to access the postgresql-simple `Connection`
which backs a Postgres `SqlBackend`.

A similar issue in the past with Sqlite was solved by the `RawSqlite`
wrapper, which is able to expose the raw connection as well as be
compatible with the wrapped `SqlBackend` that it stores.

This adds an equivalent wrapper to persistent-postgresql, which exposes
the raw connection, and therefore allows third-party libraries to access
it.