Expose orderClause from PersistQuery.hs
This is part of the followup to #1298.

This exposes `orderClause` in a similar manner to `filterClause`, to
help third-party libraries which want to format their own SQL queries
using Persistent's internal helpers.

It rewrites `orderClause` a bit to bring it in line with the other
exposed functions, but there should be no functional changes as a
result.