Add openRawSqliteConn to Database.Persist.Sqlite
This new function allows for creating a `RawSqlite SqlBackend` manually
(without exposing the `RawSqlite` constructor) for code that wants to
open such a connection without having to opt-in to the resource
management of `withRawSqliteConnInfo` and co.

This is useful in my particular use case for creating a custom pool,
since I am not constrained by the `resource-pool` API re-exposed by
`persistent-sqlite`.