chore: replace flake.nix by haskell.nix
It is by far simpler to write using haskell.nix.
It is because it automatically draws an analogy from cabal files and so on.
There is no need to be afraid of increasing dependencies if using `flake.nix`,
even if it is `default.nix` which is referenced by other nix projects.
And upgrade nix GHC version 9.6.6.